src/requirements.txt
Django~=1.9.0
psycopg2~=2.6.1
gunicorn~=19.3.0
djangorestframework~=3.3.2
celery~=3.1.19
feedparser~=5.2.1
